[SYCL][NFC] Refactor common helpers
- Dropped barely used `__SYCL_ASSERT` macro
- Moved `__SYCL_UR_ERROR_REPORT` macro and corresponding helper function
  to library
- Moved helpers thare are used in a single header to that header to
  reduce their impact on other headers
