+++ sycl/include/sycl/id.hpp
@@ -9,7 +9,6 @@
 #pragma once
 
 #include <sycl/detail/array.hpp>              // for array
-#include <sycl/detail/common.hpp>             // for InitializedVal
 #include <sycl/detail/defines.hpp>            // for __SYCL_ASSUME_INT
 #include <sycl/detail/defines_elementary.hpp> // for __SYCL_DEPRECATED, __SYCL_A...
 #include <sycl/exception.hpp> // for make_error_code, errc, exce...
